refactor: implement GetSchemaDefinition for clickhous, oceanbase, oracle, redshift and tidb
