+++ backend/plugin/schema/tidb/get_design_schema.go
@@ -1,8 +1,6 @@
 package tidb
 
 import (
-	"fmt"
-	"io"
 	"strings"
 
 	"github.com/bytebase/bytebase/backend/plugin/schema"
@@ -29,73 +27,4 @@ func GetDesignSchema(to *storepb.DatabaseSchemaMetadata) (string, error) {
 	// Make goyamlv3 happy.
 	s = strings.TrimLeft(s, "\n")
 	return s, nil
-}
-
-func writeTables(w io.StringWriter, to *storepb.DatabaseSchemaMetadata, state *databaseState) error {
-	// Follow the order of the input schemas.
-	for _, schema := range to.Schemas {
-		schemaState, ok := state.schemas[schema.Name]
-		if !ok {
-			continue
-		}
-		// Follow the order of the input tables.
-		for _, table := range schema.Tables {
-			table, ok := schemaState.tables[table.Name]
-			if !ok {
-				continue
-			}
-			if _, err := w.WriteString(getTableAnnouncement(table.name)); err != nil {
-				return err
-			}
-
-			// Avoid new line.
-			buf := &strings.Builder{}
-			if err := table.toString(buf); err != nil {
-				return err
-			}
-			if _, err := w.WriteString(buf.String()); err != nil {
-				return err
-			}
-			delete(schemaState.tables, table.name)
-		}
-	}
-	return nil
-}
-
-func writeViews(w io.StringWriter, to *storepb.DatabaseSchemaMetadata, state *databaseState) error {
-	// Follow the order of the input schemas.
-	for _, schema := range to.Schemas {
-		schemaState, ok := state.schemas[schema.Name]
-		if !ok {
-			continue
-		}
-		// Follow the order of the input views.
-		for _, view := range schema.Views {
-			view, ok := schemaState.views[view.Name]
-			if !ok {
-				continue
-			}
-			if _, err := w.WriteString(getViewAnnouncement(view.name)); err != nil {
-				return err
-			}
-
-			buf := &strings.Builder{}
-			if err := view.toString(buf); err != nil {
-				return err
-			}
-			if _, err := w.WriteString(buf.String()); err != nil {
-				return err
-			}
-			delete(schemaState.views, view.name)
-		}
-	}
-	return nil
-}
-
-func getViewAnnouncement(name string) string {
-	return fmt.Sprintf("\n--\n-- View structure for `%s`\n--\n", name)
-}
-
-func getTableAnnouncement(name string) string {
-	return fmt.Sprintf("\n--\n-- Table structure for `%s`\n--\n", name)
 }
